fix: drag item onto macrobar creates roll macro
There is a default drag and drop handler for the macrobar.  In order to
prevent the default macro from being made a hooked function needs
to return false.

An async callback that awaits does not work as it returns a Promise.
Instead Promise#then can be used to setup the macro creation
in the background and the callback can be a non-async function that
returns false when it handled the drop item.