refactor: use `Document#_updateObject` for bond changes
While using `Actor#_preUpdate` works to tranform the sheet/form changes, some
discussion in the Foundry discord makes it sound like adapting it at the sheet/form
level would be better.

This patch moves the bond conversion of `{number: bond}` to `bond[] ` into the sheet that is actually handling that form.